Updating properties, fixing warnings.

+++ src/lib3270/kybd.c
@@ -2759,21 +2759,21 @@ int kybd_prime(H3270 *hSession)
 }
 #endif /*]*/
 
-LIB3270_EXPORT int lib3270_set_unlock_delay(H3270 *session, int delay)
+LIB3270_EXPORT int lib3270_set_unlock_delay(H3270 *session, unsigned int delay)
 {
 	CHECK_SESSION_HANDLE(session);
 
-	trace("%s(%d)",__FUNCTION__,(int) delay);
+	trace("%s(%u)",__FUNCTION__,delay);
 
 	session->unlock_delay_ms = (unsigned short) delay;
 
 	return 0;
 }
 
-LIB3270_EXPORT int lib3270_get_unlock_delay(H3270 *session)
+LIB3270_EXPORT unsigned int lib3270_get_unlock_delay(H3270 *session)
 {
 	CHECK_SESSION_HANDLE(session);
-	return (int) session->unlock_delay_ms;
+	return (unsigned int) session->unlock_delay_ms;
 
 }
